* Update Indicator class * Create indicatorStatus component
